feat: Factorize send transfer logic to a reusable class
