+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/ImportFilesViewModel.kt
@@ -18,7 +18,6 @@
 package com.infomaniak.swisstransfer.ui.screen.newtransfer
 
 import android.net.Uri
-import android.util.Log
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -26,9 +25,6 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.infomaniak.core2.appintegrity.AppIntegrityManager
-import com.infomaniak.core2.appintegrity.AppIntegrityManager.Companion.APP_INTEGRITY_MANAGER_TAG
-import com.infomaniak.multiplatform_swisstransfer.SharedApiUrlCreator
 import com.infomaniak.multiplatform_swisstransfer.common.interfaces.upload.RemoteUploadFile
 import com.infomaniak.multiplatform_swisstransfer.common.interfaces.upload.UploadFileSession
 import com.infomaniak.multiplatform_swisstransfer.common.utils.mapToList
@@ -36,7 +32,6 @@ import com.infomaniak.multiplatform_swisstransfer.data.NewUploadSession
 import com.infomaniak.multiplatform_swisstransfer.managers.AppSettingsManager
 import com.infomaniak.multiplatform_swisstransfer.managers.UploadManager
 import com.infomaniak.sentry.SentryLog
-import com.infomaniak.swisstransfer.BuildConfig
 import com.infomaniak.swisstransfer.di.IoDispatcher
 import com.infomaniak.swisstransfer.ui.screen.main.settings.DownloadLimitOption
 import com.infomaniak.swisstransfer.ui.screen.main.settings.DownloadLimitOption.Companion.toTransferOption
@@ -51,31 +46,27 @@ import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.TransferOp
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.components.TransferTypeUi
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.components.TransferTypeUi.Companion.toTransferTypeUi
 import com.infomaniak.swisstransfer.ui.utils.GetSetCallbacks
-import com.infomaniak.swisstransfer.workers.UploadWorker
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.debounce
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class ImportFilesViewModel @Inject constructor(
     private val appSettingsManager: AppSettingsManager,
-    private val appIntegrityManager: AppIntegrityManager,
     private val savedStateHandle: SavedStateHandle,
     private val importationFilesManager: ImportationFilesManager,
-    private val sharedApiUrlCreator: SharedApiUrlCreator,
     private val uploadManager: UploadManager,
-    private val uploadWorkerScheduler: UploadWorker.Scheduler,
+    private val transferSendManager: TransferSendManager,
     @IoDispatcher private val ioDispatcher: CoroutineDispatcher,
 ) : ViewModel() {
 
-    private val _sendActionResult = MutableStateFlow<SendActionResult?>(SendActionResult.NotStarted)
-    val sendActionResult = _sendActionResult.asStateFlow()
-
-    private val _integrityCheckResult = MutableStateFlow(AppIntegrityResult.Idle)
-    val integrityCheckResult = _integrityCheckResult.asStateFlow()
+    val sendActionResult by transferSendManager::sendActionResult
+    val integrityCheckResult by transferSendManager::integrityCheckResult
 
     @OptIn(FlowPreview::class)
     val importedFilesDebounced = importationFilesManager.importedFiles
@@ -138,82 +129,19 @@ class ImportFilesViewModel @Inject constructor(
         }
     }
 
-    private fun sendTransfer(attestationToken: String) {
-        _sendActionResult.update { SendActionResult.Pending }
+    fun sendTransfer() {
         viewModelScope.launch(ioDispatcher) {
-            runCatching {
-                val uuid = uploadManager.createAndGetUpload(generateNewUploadSession()).uuid
-                uploadManager.initUploadSession(
-                    attestationHeaderName = AppIntegrityManager.ATTESTATION_TOKEN_HEADER,
-                    attestationToken = attestationToken,
-                )!! // TODO: Handle ContainerErrorsException here
-                uploadWorkerScheduler.scheduleWork(uuid)
-                _sendActionResult.update {
-                    val totalSize = importationFilesManager.importedFiles.value.sumOf { it.fileSize }
-                    SendActionResult.Success(totalSize)
-                }
-            }.onFailure { exception ->
-                SentryLog.e(TAG, "Failed to start the upload", exception)
-                _sendActionResult.update { SendActionResult.Failure }
-            }
+            transferSendManager.sendTransfer(generateNewUploadSession())
         }
     }
 
     fun resetSendActionResult() {
-        _sendActionResult.value = SendActionResult.NotStarted
-    }
-
-    //region App Integrity
-    fun checkAppIntegrity() {
-        _integrityCheckResult.value = AppIntegrityResult.Ongoing
-        viewModelScope.launch(ioDispatcher) {
-            runCatching {
-                appIntegrityManager.getChallenge(
-                    onSuccess = { requestAppIntegrityToken(appIntegrityManager) },
-                    onFailure = ::setFailedIntegrityResult,
-                )
-            }.onFailure { exception ->
-                SentryLog.e(TAG, "Failed to start the upload", exception)
-                _sendActionResult.update { SendActionResult.Failure }
-            }
-        }
-    }
-
-    private fun requestAppIntegrityToken(appIntegrityManager: AppIntegrityManager) {
-        appIntegrityManager.requestClassicIntegrityVerdictToken(
-            onSuccess = { token ->
-                SentryLog.i(APP_INTEGRITY_MANAGER_TAG, "request for app integrity token successful $token")
-                getApiIntegrityVerdict(appIntegrityManager, token)
-            },
-            onFailure = ::setFailedIntegrityResult,
-        )
-    }
-
-    private fun getApiIntegrityVerdict(appIntegrityManager: AppIntegrityManager, appIntegrityToken: String) {
-        viewModelScope.launch(ioDispatcher) {
-            appIntegrityManager.getApiIntegrityVerdict(
-                integrityToken = appIntegrityToken,
-                packageName = BuildConfig.APPLICATION_ID,
-                targetUrl = sharedApiUrlCreator.createUploadContainerUrl,
-                onSuccess = { attestationToken ->
-                    SentryLog.i(APP_INTEGRITY_MANAGER_TAG, "Api verdict check")
-                    Log.i(APP_INTEGRITY_MANAGER_TAG, "getApiIntegrityVerdict: $attestationToken")
-                    _integrityCheckResult.value = AppIntegrityResult.Success
-                    sendTransfer(attestationToken)
-                },
-                onFailure = ::setFailedIntegrityResult,
-            )
-        }
-    }
-
-    private fun setFailedIntegrityResult() {
-        _integrityCheckResult.value = AppIntegrityResult.Fail
+        transferSendManager.resetSendActionResult()
     }
 
     fun resetIntegrityCheckResult() {
-        _integrityCheckResult.value = AppIntegrityResult.Idle
+        transferSendManager.resetIntegrityCheckResult()
     }
-    //endregion
 
     private suspend fun removeOldData() {
         importationFilesManager.removeLocalCopyFolder()
