+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/importfiles/ImportFilesScreen.kt
@@ -144,7 +144,7 @@ fun ImportFilesScreen(
         addFiles = importFilesViewModel::importFiles,
         closeActivity = closeActivity,
         integrityCheckResult = { integrityCheckResult },
-        checkAppIntegrity = importFilesViewModel::checkAppIntegrity,
+        sendTransfer = importFilesViewModel::sendTransfer,
         shouldStartByPromptingUserForFiles = true,
         isTransferStarted = { sendActionResult != SendActionResult.NotStarted },
         snackbarHostState = snackbarHostState,
@@ -202,7 +202,7 @@ private fun ImportFilesScreen(
     closeActivity: () -> Unit,
     shouldStartByPromptingUserForFiles: Boolean,
     integrityCheckResult: () -> AppIntegrityResult,
-    checkAppIntegrity: () -> Unit,
+    sendTransfer: () -> Unit,
     isTransferStarted: () -> Boolean,
     snackbarHostState: SnackbarHostState? = null,
 ) {
@@ -226,7 +226,7 @@ private fun ImportFilesScreen(
                 shouldShowEmailAddressesFields = { shouldShowEmailAddressesFields },
                 transferAuthorEmail = transferAuthorEmail,
                 integrityCheckResult = integrityCheckResult,
-                checkAppIntegrityThenSendTransfer = checkAppIntegrity,
+                sendTransfer = sendTransfer,
                 isTransferStarted = isTransferStarted,
             )
         },
@@ -406,7 +406,7 @@ private fun SendButton(
     shouldShowEmailAddressesFields: () -> Boolean,
     transferAuthorEmail: GetSetCallbacks<String>,
     integrityCheckResult: () -> AppIntegrityResult,
-    checkAppIntegrityThenSendTransfer: () -> Unit,
+    sendTransfer: () -> Unit,
     isTransferStarted: () -> Boolean,
 ) {
     val remainingFilesCount = filesToImportCount()
@@ -431,7 +431,7 @@ private fun SendButton(
         showIndeterminateProgress = { integrityCheckResult() == AppIntegrityResult.Ongoing || isTransferStarted() },
         enabled = { importedFiles().isNotEmpty() && !isImporting && isSenderEmailCorrect && !isTransferStarted() },
         progress = progress,
-        onClick = { checkAppIntegrityThenSendTransfer() },
+        onClick = { sendTransfer() },
     )
 }
 
@@ -498,7 +498,7 @@ private fun Preview(@PreviewParameter(FileUiListPreviewParameter::class) files: 
             closeActivity = {},
             shouldStartByPromptingUserForFiles = false,
             integrityCheckResult = { AppIntegrityResult.Idle },
-            checkAppIntegrity = {},
+            sendTransfer = {},
             isTransferStarted = { false },
         )
     }
